BARK-80: Move VERSION into lib/version
