+++ src/lib/version.ts
